Extract new class 'ViewMenu' from TripleAMenu and BasicGameMenuBar
